requirements.txt
google-genai
pymongo
python-dotenv
requests
deep-translator